+++ www/src/utils/colors.js
@@ -58,13 +58,14 @@ const colors = {
   // original palette by @SachaG
   // @see https://www.figma.com/file/J6IYJEtdRmwJQOrcZ2DfvxDD/Gatsby
   gatsby: `#663399`, // was #744c9e
-  lilac: `#9D7CBF`,
+  lilac: `#8c65b3`,
   // accent color from the "bolder palette" by @ArchieHicklin
   // @see https://github.com/gatsbyjs/gatsby/issues/1173#issuecomment-309415650
   accent: `#ffb238`, // "Mustard",
   success: `#37b635`,
   warning: `#ec1818`,
   ui: {
+    border: `#ede7f3`,
     bright: `#e0d6eb`,
     light: `#f5f3f7`,
     whisper: `#fbfafc`,
@@ -75,6 +76,10 @@ const colors = {
     calm: gray(46, 270),
     bright: gray(64, 270),
     light: gray(80, 270),
+  },
+  code: {
+    remove: `#e45c5c`,
+    add: `#4a9c59`,
   },
 }